+++ semcore/d3-chart/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [3.50.0] - 2024-06-28
+
+### Added
+
+- `CompactHorizontalBar` chart.
+
 ## [3.49.0] - 2024-07-13
 
 ### Changed